+++ gui/settings/settings.gd
@@ -1,7 +1,9 @@
 extends PanelContainer
 class_name Settings
 
-@export var gui: GUI
+@export var previous_menu: Control
+@export var input_mapper: InputMapper
+
 @onready var selection_container: SelectionContainer = $MarginContainer/SelectionContainer
 @onready var volume_button: VolumeButton = $MarginContainer/SelectionContainer/VolumeButton
 @onready var fullscreen_button: Button = $MarginContainer/SelectionContainer/FullscreenButton
@@ -11,16 +13,16 @@ func _ready() -> void:
 
 func _on_close_button_pressed() -> void:
 	hide()
-	gui.menu.show()
-	gui.menu.selection_container.get_buttons()[0].grab_focus()
+	previous_menu.show()
+	previous_menu.selection_container.get_buttons()[0].grab_focus()
 
 	# Save the volume settings
 	SaveManager.set_volume(volume_button.volume)
 
 func _on_input_button_pressed() -> void:
 	hide()
-	gui.input_mapper.show()
-	gui.input_mapper.selection_container.get_buttons()[0].grab_focus()
+	input_mapper.show()
+	input_mapper.selection_container.get_buttons()[0].grab_focus()
 
 func _on_fullscreen_button_pressed() -> void:
 	Utils.toggle_fullscreen()