+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # CHANGELOG
 
 ## v0.1.4
-- Added more animation to the player (assets/images/player.png)
+- Added more animation to the player ("res://assets/images/player.png")
     - Separated the animation part from the propagation mechaninc "handle_animation"
 - In "Utils.gd" added "toggle_fullscreen" method
 - In "Settings" added "FullScreenButton"
+- Added "ChargerStation" to charge the player stamina (now the stamina will not auto-refill)
+- Added "HealthBar"
 
 ## v0.1.3
 - Added a "DeathZone" node to "Wall" scene, now when the player is stuck in a wall the game is over.